+++ src/index.tsx
@@ -68,36 +68,36 @@ const Main = () => {
         <div className="d-flex h-100">
             <div className="floating-card">
                 <div className="title">
-                    {updateAvailable ? 'Update' : 'Install'}
+                    Bluetooth Low Energy Application{' '}
+                    {updateAvailable ? 'Update' : 'Installation'}
                 </div>
                 <div className="content p-3 pt-4">
                     {progress === undefined && (
                         <>
                             {updateAvailable && (
                                 <>
-                                    <p>An update is available</p>
-                                    <p className="mb-0">
-                                        Installed version: {version}.
+                                    <p>
+                                        An update is available for this
+                                        application
                                     </p>
-                                    <p>Available version: {bleVersion}.</p>
+                                    <ul>
+                                        <li>Installed version: {version}.</li>
+                                        <li>
+                                            Available version: {bleVersion}.
+                                        </li>
+                                    </ul>
                                 </>
                             )}
 
-                            <p>
-                                {!exePath && progress === undefined && (
-                                    <>
-                                        Bluetooth Low Energy is now a standalone
-                                        app.
-                                        {process.platform === 'win32' && (
-                                            <p className="mt-1">
-                                                The standalone application can
-                                                be uninstalled from the control
-                                                panel.
-                                            </p>
-                                        )}
-                                    </>
-                                )}
-                            </p>
+                            {!exePath && progress === undefined && (
+                                <p>
+                                    The Bluetooth Low Energy application has
+                                    been converted to a standalone application
+                                    for compatibility reasons. It may however
+                                    still be opened from nRF Connect for
+                                    Desktop.
+                                </p>
+                            )}
 
                             {exePath &&
                                 dirname(exePath) !== programDirectory() && (
@@ -124,25 +124,31 @@ const Main = () => {
                                 <>
                                     <div className="d-flex flex-column mb-0">
                                         <p>
-                                            The standalone application will be
-                                            installed at:
+                                            Click the button below to install
+                                            the application at the following
+                                            path:
                                         </p>
                                         <p className="path">
                                             {programDirectory()}
                                         </p>
-                                        <p className="mb-4">
-                                            File size: {downloadSize}MB
-                                        </p>
+                                        <p>File size: {downloadSize}MB</p>
+                                        {process.platform === 'win32' && (
+                                            <p>
+                                                The standalone application may
+                                                be uninstalled from the Windows
+                                                Control Panel.
+                                            </p>
+                                        )}
                                     </div>
 
-                                    <div className="d-flex justify-content-end">
+                                    <div className="d-flex justify-content-end mt-4">
                                         {exePath && progress === undefined && (
                                             <button
                                                 type="button"
                                                 className="btn btn-secondary font-size-12 mr-1"
                                                 onClick={runExecutable}
                                             >
-                                                Open current version
+                                                Open Current Version
                                             </button>
                                         )}
                                         {progress === undefined && (
@@ -154,7 +160,7 @@ const Main = () => {
                                             >
                                                 {updateAvailable
                                                     ? 'Update Application'
-                                                    : 'Download and install'}
+                                                    : 'Download and Install'}
                                             </button>
                                         )}
                                     </div>
@@ -164,14 +170,13 @@ const Main = () => {
                     )}
                     {progress !== undefined && (
                         <>
-                            <p>Downloading...</p>
+                            <p>Downloading {exePath ? 'update' : 'files'}...</p>
                             <ProgressBar now={progress} />
                             <div className="d-flex justify-content-end mt-3">
                                 <button
                                     type="button"
-                                    className="btn btn-outline-dark font-size-12"
+                                    className="btn btn-secondary font-size-12"
                                     onClick={cancel}
-                                    style={{ width: '150px' }}
                                 >
                                     Cancel
                                 </button>